Main functions

API Reference

ProxAL.ProxALEvaluator — Type

ProxALEvaluator(
    case_file::String,
    load_file::String,
    modelinfo::ModelInfo,
    algparams::AlgParams,
    space::AbstractBackend = JuMPBackend(),
    comm::Union{MPI.Comm,Nothing} = MPI.COMM_WORLD
)

Instantiate multi-period ACOPF specified in case_file with loads in load_file with model parameters modelinfo, algorithm parameters algparams, modeling backend space, and a MPI communicator comm.

ProxAL.NonDecomposedModel — Type

NonDecomposedModel(
    case_file::String,
    load_file::String,
    modelinfo::ModelInfo,
    algparams::AlgParams,
    space::AbstractBackend=JuMPBackend(),
    time_horizon_start = 1,
)

Instantiate non-decomposed multi-period ACOPF instance specified in case_file with loads in load_file with model parameters modelinfo and algorithm parameters algparams. The problem is defined over the horizon [time_horizon_start, modelinfo.num_time_periods]

ProxAL.optimize! — Method

optimize!(nlp::ProxALEvaluator)

Solve problem using the nlp evaluator of the decomposition algorithm.

ProxAL.optimize! — Method

optimize!(nlp::NonDecomposedModel)

Solve problem using the nlp evaluator of the nondecomposed model.