MPI communication

API reference

ProxAL.comm_max — Method

comm_max(data, comm)

Collective to reduce and return the maximum over scalar data.

ProxAL.comm_neighbors! — Method

comm_neighbors!(data, blocks, runinfo, pattern, comm)

Nonblocking communication with a given pattern. An array of requests is returned.

ProxAL.comm_sum! — Method

comm_sum!(data, comm)

Collective to reduce the sum over array data.

ProxAL.comm_sum — Method

comm_sum(data::Float64, comm)

Collective to reduce and return the sum over scalar data.

ProxAL.comm_wait! — Method

comm_wait!(requests)

Wait until the communciation requests requests have been fulfilled.

ProxAL.is_comm_pattern — Method

is_comm_pattern(t, tn, k, kn, pattern)

Do period t, tn and contingencies k, kn match the communication pattern?

ProxAL.is_my_work — Method

is_my_work(blk, comm)

Returns a boolean whether the block blk is assigned to this rank.

ProxAL.whoswork — Method

whoswork(blk, comm)

To which rank is block blk currently assigned to.