Optimal Power Flow

Data parsing

ProxAL.parse_file — Function

parse_file(datafile::String)

Parse MATPOWER or PSSE instances using ExaPF's parsers. Return full dataset as Dict{String, Array{Float64, 2}, with entries

"bus": specifications for all buses in the network

"branch": specifications for all branches in the network

"gen": specifications for all generators in the network

"costs": costs coefficients.

"baseMVA": baseMVA of the network

ProxAL.RawData — Type

RawData

Specifies the ACOPF instance data.

baseMVA: imported with ExaPF parser

bus_arr: imported with ExaPF parser

branch_arr: imported with ExaPF parser

gen_arr: imported with ExaPF parser

costgen_arr: imported with ExaPF parser

pd_arr: read from .Pd file

qd_arr: read from .Qd file

ctgs_arr: read from .Ctgs file

ProxAL.opf_loaddata — Function

opf_loaddata(
    raw::RawData;
    time_horizon_start::Int=1,
    time_horizon_end::Int=0,
    load_scale::Float64=1.0,
    ramp_scale::Float64=0.0,
    corr_scale::Float64=0.1,
    lineOff=Line()
)

Loads the multi-period ACOPF instance data from raw with the time horizon defined to be [time_horizon_start, time_horizon_end]. Note that time_horizon_end = 0 indicates as many as possible (the number of columns in raw.pd_arr).

All loads in all time periods will be multiplied by load_scale. The ramp_scale is the factor which multiplies $p_{g}^{max}$ to get generator ramping $r_g$. The corr_scale is the factor which multiplies $r_g$ to get generator ramping for corrective control. These are set in ModelInfo. See Model parameters.

lineOff is a transmission line that can be deleted to represent a contingency.